sound: Use a 32-bit buffer for audio samples
While testing MIDI files, it was found in rare cases that the 16-bit
audio samples could overflow. When this happened, the sample would flip
sign and introduce pops and crackle. This is due to the sudden jump
between near maximum and minimum values.

One option to prevent this would have been to reduce the overall volume
of the ym2413 chips. But this would have affected more than just the
loudest MIDI files that were showing th e problem. Another possibility
would have been investigating floating-point audio samples.

However, for now the chosen solution is to add some headroom by summing
the various audio chips into a 32-bit buffer. Once the samples are all
added up, the range is limited to the 16-bit range before passing them
to the sound card.

This means that, instead of the sample wrapping around and generating
pops, the peak amplitude of the wave is simply flattened at the limit.
This has not been noticed when listening to the same files that would
previously show pops / crackle.

+++ source/colecovision.c
@@ -41,7 +41,7 @@ static void     colecovision_update_settings (void *context_ptr);
 /*
  * Callback to supply audio frames.
  */
-static void colecovision_audio_callback (void *context_ptr, int16_t *stream, uint32_t count)
+static void colecovision_audio_callback (void *context_ptr, int32_t *stream, uint32_t count)
 {
     ColecoVision_Context *context = (ColecoVision_Context *) context_ptr;
