+++ source/sms.c
@@ -63,7 +63,7 @@ static void        sms_update_settings (void *context_ptr);
 /*
  * Callback to supply audio frames.
  */
-static void sms_audio_callback (void *context_ptr, int16_t *stream, uint32_t count)
+static void sms_audio_callback (void *context_ptr, int32_t *stream, uint32_t count)
 {
     SMS_Context *context = (SMS_Context *) context_ptr;
